change/refactor: remove a few more utils star imports
