+++ CreeDictionary/DatabaseManager/xml_importer.py
@@ -2,14 +2,7 @@ import re
 import xml.etree.ElementTree as ET
 from collections import defaultdict
 from pathlib import Path
-from typing import (
-    DefaultDict,
-    Dict,
-    List,
-    NamedTuple,
-    Set,
-    Tuple,
-)
+from typing import DefaultDict, Dict, List, NamedTuple, Set, Tuple
 
 from colorama import init
 
@@ -20,15 +13,12 @@ from DatabaseManager.log import DatabaseManagerLogger
 from DatabaseManager.xml_consistency_checker import (
     does_inflectional_category_match_xml_entry,
 )
-from utils import (
-    PartOfSpeech,
-    fst_analysis_parser,
-    XMLEntry,
-)
+from utils import PartOfSpeech, fst_analysis_parser
 from utils.crkeng_xml_utils import (
-    convert_xml_inflectional_category_to_word_class,
     IndexedXML,
+    convert_xml_inflectional_category_to_word_class,
 )
+from utils.data_classes import XMLEntry
 from utils.profiling import timed
 
 init()  # for windows compatibility
