+++ CreeDictionary/utils/__init__.py
@@ -1,6 +1,4 @@
 from .cree_lev_dist import get_modified_distance
-from .data_classes import *
 from .enums import *
-from .fst_analysis_parser import extract_lemma_text_and_word_class, extract_word_class
 from .shared_res_dir import shared_res_dir
 from .vars import *
